refactor(InputView, Car): Car에 있는 자동차 이름 예외 처리 코드 InputView 클래스로 이동
